Misc.
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@48919 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/recsrc/IndexTableScan.cpp
@@ -48,7 +48,7 @@ IndexTableScan::IndexTableScan(CompilerScratch* csb, const string& name, UCHAR s
 {
 	fb_assert(m_index);
 
-	size_t size = sizeof(Impure) + 2u * static_cast<size_t>(m_length);
+	size_t size = sizeof(Impure) + 2u * m_length;
 	size = FB_ALIGN(size, FB_ALIGNMENT);
 	m_offset = size;
 	size += sizeof(index_desc);
